grpc-app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
protobuf==5.27.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
setuptools==70.2.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlmodel==0.0.19
starlette==0.37.2
style==1.1.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
update==0.0.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
